zakaria_atyq_NLP_analyse_sentiment/Analyse de Sentiments avec BERT/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch==2.4.1
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12